+++ src/studio/lib/provider_http_client_io.dart
@@ -1,3 +1,5 @@
 import 'package:http/http.dart' as http;
 
 http.Client createProviderHttpClientImpl() => http.Client();
+
+http.Client createPublicObjectHttpClientImpl() => http.Client();