+++ src/studio/lib/provider_http_client_web.dart
@@ -4,3 +4,5 @@ import 'package:http/http.dart' as http;
 http.Client createProviderHttpClientImpl() {
   return BrowserClient()..withCredentials = true;
 }
+
+http.Client createPublicObjectHttpClientImpl() => BrowserClient();